refactor(#792): extract ChainRuntimeState tap methods into their own file (sweep)
runtime_state.rs was 721 lines mixing the ChainRuntimeState struct + control
(volume/mute/draining), DI-loop, diagnostics, and the tap-subscription methods.
Moved the tap concern — subscribe/prune input+stream taps, adopt_taps_from,
stream_count, input_routing_for_stream, poll_stream — into
runtime_state_taps.rs (a sibling `impl ChainRuntimeState`; all fields are
pub(crate), so it's a pure move with the same public API). runtime_state.rs
drops to 520 and leaves the debt list. Engine tests pass, zero warnings.

God-file sweep note: the biggest infra-cpal targets (controller.rs 1077,
stream_builder.rs, dsp_worker.rs, jack_supervisor) are heavily #[cfg(linux+jack)]
— can't be compile-verified on macOS, so they need a linux build to split safely.
