+++ scripts/validate.sh
@@ -71,7 +71,6 @@ crates/engine/src/runtime_audio_frame.rs
 crates/engine/src/runtime_block_builders.rs
 crates/engine/src/runtime_dsp.rs
 crates/engine/src/runtime_graph.rs
-crates/engine/src/runtime_state.rs
 crates/engine/src/runtime_tests.rs
 crates/engine/src/runtime.rs
 crates/engine/src/stream_isolation_tests.rs
